gun: move Gun_FindTargetPoint to libtrx
