+++ src/libtrx/game/gun/misc.c
@@ -1,0 +1,19 @@
+#include "game/gun/misc.h"
+
+#include "game/const.h"
+#include "game/items.h"
+#include "game/math.h"
+
+void Gun_FindTargetPoint(const ITEM *const item, GAME_VECTOR *const target)
+{
+    const BOUNDS_16 *const bounds = &Item_GetBestFrame(item)->bounds;
+    const int32_t x = bounds->min.x + (bounds->max.x - bounds->min.x) / 2;
+    const int32_t y = bounds->min.y + (bounds->max.y - bounds->min.y) / 3;
+    const int32_t z = bounds->min.z + (bounds->max.z - bounds->min.z) / 2;
+    const int32_t cy = Math_Cos(item->rot.y);
+    const int32_t sy = Math_Sin(item->rot.y);
+    target->pos.x = item->pos.x + ((cy * x + sy * z) >> W2V_SHIFT);
+    target->pos.y = item->pos.y + y;
+    target->pos.z = item->pos.z + ((cy * z - sy * x) >> W2V_SHIFT);
+    target->room_num = item->room_num;
+}
