+++ src/libtrx/include/libtrx/game/gun/misc.h
@@ -3,4 +3,4 @@
 #include "../items/types.h"
 #include "../types.h"
 
-extern void Gun_FindTargetPoint(const ITEM *item, GAME_VECTOR *target);
+void Gun_FindTargetPoint(const ITEM *item, GAME_VECTOR *target);
